Merge major architectural changes

See merge request COMP2511/20T3/m17a-and_i_oop!9

+++ src/unsw/gloriaromanus/component/VeryHighTax.java
@@ -3,7 +3,7 @@ package unsw.gloriaromanus.component;
 public class VeryHighTax implements TaxLevel {
 
     @Override
-    public float getTaxRate() {
+    public int getTaxRate() {
         return 25;
     }
